requirements.txt
loguru~=0.7.2
pika~=1.3.2
PyYAML~=6.0.1
oss2~=2.18.2
requests~=2.25.1
websockets~=12.0
pytest~=7.4.2
packaging~=23.2
Pillow~=10.1.0